refactor(DTO): MoveDTO -> MoveRequest domain 로직으로 이전
